加header footer 調整樣式
